# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.UpdateNovelPayload:
    properties:
      author:
        maxLength: 255
        type: string
      genre:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
...
  store.Chapter:
    properties:
      chapter_number:
        type: number
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_locked:
        type: boolean
      is_read:
        type: boolean
      novel_id:
        type: integer
      price:
        type: integer
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  store.Novel:
    properties:
      author:
        type: string
      chapters:
        items:
          $ref: '#/definitions/store.Chapter'
        type: array
      created_at:
        type: string
      genre:
        items:
          type: string
        type: array
      id:
        type: integer
      image_url:
        type: string
      synopsis:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  swagger.EnvelopeError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for an online novel platform — read, publish, and manage
    digital stories.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Govel API
paths:
...
  /novels:
    post:
      consumes:
      - multipart/form-data
      description: Create a new novel with title, author, synopsis, genre, and optional
        image. Admon only
      parameters:
      - description: Novel Title
        in: formData
        name: title
        required: true
        type: string
      - description: Author of the Novel
        in: formData
        name: author
        required: true
        type: string
      - description: Synopsis of the Novel
        in: formData
        name: synopsis
        required: true
        type: string
      - description: Genre of the Novel
        in: formData
        name: genre
        required: true
        type: string
      - description: Cover image of the Novel
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Novel created successfully
          schema:
            $ref: '#/definitions/store.Novel'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Create a new novel
      tags:
      - novels
  /novels/{novelID}:
    delete:
      consumes:
      - application/json
      description: Delete novel by ID. Admin only
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Delete novel succesfully
          schema:
            type: ""
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete novel
      tags:
      - novels
    get:
      description: Get detailed information about a specific novel by its ID, including
        chapters
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detail novel with chapters
          schema:
            $ref: '#/definitions/store.Novel'
        "400":
          description: Invalid novel ID
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      summary: Get novel detail
      tags:
      - novels
    patch:
      consumes:
      - application/json
      description: Update an existing novel's title, author, synopsis, or genre. Admin
        only.
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.UpdateNovelPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Updated novel
          schema:
            $ref: '#/definitions/store.Novel'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Update novel
      tags:
      - novels
...
  /novels/{novelID}/image:
    patch:
      consumes:
      - multipart/form-data
      description: Update the cover image of a novel. Admin only.
      parameters:
      - description: Novel ID
        in: path
        name: novelID
        required: true
        type: integer
      - description: Image file (jpg, png, etc.)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated novel with new image
          schema:
            $ref: '#/definitions/store.Novel'
        "400":
          description: Bad request (e.g. no image)
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "401":
          description: Unauthorize
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "404":
          description: Novel not found
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/swagger.EnvelopeError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Change novel image
      tags:
      - novels
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"